extract resolve and normalize into own class
